Refactor: Split qr-styler.ts into smaller modules
Refactored src/utils/qr-styler.ts into smaller, more manageable files to improve code organization and readability. No functionality changes were introduced. Unused imports and files were removed.
